RF: Split train UI from its execution for ease of accessing its documentation
